Refactor WorkOrderDetails component
Refactor WorkOrderDetails.tsx into custom hooks and UI components for improved maintainability and testability.
