+++ src/utils/workOrderHelpers.ts
@@ -1,0 +1,37 @@
+export const getPriorityColor = (priority: string) => {
+  switch (priority) {
+    case 'high':
+      return 'bg-red-100 text-red-800 border-red-200';
+    case 'medium':
+      return 'bg-yellow-100 text-yellow-800 border-yellow-200';
+    case 'low':
+      return 'bg-green-100 text-green-800 border-green-200';
+    default:
+      return 'bg-gray-100 text-gray-800 border-gray-200';
+  }
+};
+
+export const getStatusColor = (status: string) => {
+  switch (status) {
+    case 'submitted':
+      return 'bg-blue-100 text-blue-800 border-blue-200';
+    case 'accepted':
+      return 'bg-purple-100 text-purple-800 border-purple-200';
+    case 'assigned':
+      return 'bg-orange-100 text-orange-800 border-orange-200';
+    case 'in_progress':
+      return 'bg-yellow-100 text-yellow-800 border-yellow-200';
+    case 'on_hold':
+      return 'bg-gray-100 text-gray-800 border-gray-200';
+    case 'completed':
+      return 'bg-green-100 text-green-800 border-green-200';
+    case 'cancelled':
+      return 'bg-red-100 text-red-800 border-red-200';
+    default:
+      return 'bg-gray-100 text-gray-800 border-gray-200';
+  }
+};
+
+export const formatStatus = (status: string) => {
+  return status.replace(/_/g, ' ').replace(/\b\w/g, l => l.toUpperCase());
+};
